+++ lib/src/quantity/number.dart
@@ -19,3 +19,4 @@ export 'src/number/real.dart';
 export 'src/number/util/converters.dart';
 export 'src/number/util/erf.dart';
 export 'src/number/util/jenkins_hash.dart';
+export 'src/number/util/romans_exception.dart';